Create fv3core python package (#51)
This PR creates a package `fv3core` which can be installed using a `setup.py` file.

Translate classes and test infrastructure is moved to the `tests` directory for the package. This includes files previously under `fv3/tests` and `fv3/translate`. All other files in `fv3` have been moved into a `fv3core` directory, used as the python module.

For now, I've imported the two routines we want to use from outside the package (in fv3gfs-python) to the top-level of the package: `fv_dynamics` and `fv_subgridz`. This will likely change when we have a discussion about design and refactors.

The main "new" or substantially changed files to look at are:

- setup.py
- I added a HISTORY.md but it is blank for now
- docker/Dockerfile
- renamed `test_python_modules.py` to `test_translate.py`
- mounts for dev endpoints in Makefile are now controlled by a `DEV_MOUNTS` variable