Move decorators out of gt4py_utils (#136)
I'm about to add significant features to the gt4py decorators, and don't want to put this logic in gt4py_utils.py which already has a lot of logic. This PR moves the stencil logic out to its own module.

Part of refactoring this meant changing the way "external" code (tests and examples) change the fv3core backend. I added `set_backend` and `set_rebuild` functions at the top level which allow the internal code to be refactored without breaking external code. This will greatly reduce the number of places code needs to be updated when we modify the way backend is set and stored in upcoming refactors.