Add python-to-disk serialization for stencils (#147)
This PR adds two kinds of serialization for stencils, controlled by functions `fv3core.enable_stencil_report` and `fv3core.disable_stencil_report`. When used, the model will save numpy arrays (for storages or numpy arrays) and/or a report of the min, max, mean, standard deviation, and md5sum of the data.